pyproject.toml
[project]
name = "ai-video-generator"
version = "0.1.0"
description = "AI-powered video generator with script enhancement, voiceovers, and captions"
authors = [{name = "Jeff Jackson", email = "jeff99jackson99@users.noreply.github.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "moviepy>=1.0.3",
    "pydub>=0.25.1",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.8.0",
    "openai-whisper>=20231117",
    "TTS>=0.22.0",
    "gTTS>=2.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
